Add ticket create operation

+++ cs-board/app/Http/Controllers/TicketController.php
@@ -23,6 +23,8 @@ class TicketController extends Controller
      */
     public function create()
     {
+        $this->authorize('create ticket');
+
         return view('tickets.create');
     }
 
@@ -31,8 +33,10 @@ class TicketController extends Controller
      */
     public function store(StoreTicketRequest $request)
     {
+        $this->authorize('create ticket');
+        //dd($request->validated());
         Ticket::create($request->validated());
- 
+
         return redirect()->route('tickets.index');
     }
 
@@ -49,6 +53,8 @@ class TicketController extends Controller
      */
     public function edit(Ticket $ticket)
     {
+        $this->authorize('edit ticket');
+
         return view('tickets.edit', compact('ticket'));
     }
 
@@ -57,6 +63,8 @@ class TicketController extends Controller
      */
     public function update(UpdateTicketRequest $request, Ticket $ticket)
     {
+        $this->authorize('edit ticket');
+
         $ticket->update($request->validated());
  
         return redirect()->route('tickets.index');
@@ -67,6 +75,8 @@ class TicketController extends Controller
      */
     public function destroy(Ticket $ticket)
     {
+        $this->authorize('delete ticket');
+
         $ticket->delete();
  
         return redirect()->route('tickets.index');
